+++ index.html
@@ -11,13 +11,5 @@
     <div id="root"></div>
     <h3>Namaste Everyone using HTML</h3>
   </body>
-  <script
-    crossorigin
-    src="https://unpkg.com/react@18/umd/react.development.js"
-  ></script>
-  <script
-    crossorigin
-    src="https://unpkg.com/react-dom@18/umd/react-dom.development.js"
-  ></script>
-  <script src="App.js"></script>
+  <script type="module" src="App.js"></script>
 </html>
